+++ modules/03_network.ps1
@@ -1,5 +1,5 @@
 ﻿###############################################################################
-# Maintenance Toolkit 3.7.0 - Modulo report rete
+# Maintenance Toolkit 3.7.1 - Modulo report rete
 ###############################################################################
 
 . "$PSScriptRoot\00_common.ps1"